+++ src/codes/golomb.rs
@@ -23,7 +23,7 @@
 //!
 //! The supported range is [0 . . 2⁶⁴) for *b* in [1 . . 2⁶⁴), but writing
 //! 2⁶⁴ – 1 when *b* = 1 requires writing the unary code for 2⁶⁴ – 1, which
-//! might not be possible depending on the [`BitWrite`](crate::traits::BitWrite)
+//! might not be possible depending on the [`BitWrite`]
 //! implementation (and would require writing 2⁶⁴ bits anyway).
 //!
 //! # References