+++ src/impls/mem_word_reader.rs
@@ -151,7 +151,7 @@ impl<W: Word, B: AsRef<[W]>> WordSeek for MemWordReader<W, B, false> {
     fn set_word_pos(&mut self, word_index: u64) -> Result<(), WordError> {
         if word_index > self.data.as_ref().len() as u64 {
             Err(WordError::UnexpectedEof {
-                word_pos: self.word_index,
+                word_pos: word_index as usize,
             })
         } else {
             self.word_index = word_index as usize;